+++ conf/atos_bologna.ini
@@ -309,7 +309,7 @@ bundle_src_dir      = $SCRATCH/bundle_cache
 # FORECASTS
 # =========
 
-[canonical_forecasts]
+[canonical_forecast_arpege_global798c22]
 prefix              = canonical
 # profile (ntasks = ntasks/node)
 time                = 01:00:00
@@ -317,7 +317,15 @@ ntasks              = 32
 nnodes              = 4
 openmp              = 4
 
-[standalone_forecasts]
+[canonical_forecast_arpege_apl_global798c22]
+prefix              = canonical
+# profile (ntasks = ntasks/node)
+time                = 01:00:00
+ntasks              = 32
+nnodes              = 4
+openmp              = 4
+
+[standalone_forecast_alaro0_antwrp1300]
 prefix              = standalone
 pgd_source          = static
 surf_ic_source      = static
@@ -326,6 +334,57 @@ time                = 01:00:00
 ntasks              = 32
 nnodes              = 1
 openmp              = 4
+
+[standalone_forecast_alaro1_antwrp1300]
+prefix              = standalone
+pgd_source          = static
+surf_ic_source      = static
+# profile (ntasks = ntasks/node)
+time                = 01:00:00
+ntasks              = 32
+nnodes              = 1
+openmp              = 4
+
+[standalone_forecast_arome_corsica2500]
+prefix              = standalone
+pgd_source          = static
+surf_ic_source      = static
+# profile (ntasks = ntasks/node)
+time                = 01:00:00
+ntasks              = 32
+nnodes              = 1
+openmp              = 4
+
+[standalone_forecast_arpege_globaltst149c24]
+prefix              = standalone
+pgd_source          = static
+surf_ic_source      = static
+# profile (ntasks = ntasks/node)
+time                = 01:00:00
+ntasks              = 32
+nnodes              = 1
+openmp              = 4
+
+[standalone_forecast_arpege_nosfx_globaltst149c24]
+prefix              = standalone
+pgd_source          = static
+surf_ic_source      = static
+# profile (ntasks = ntasks/node)
+time                = 01:00:00
+ntasks              = 32
+nnodes              = 1
+openmp              = 4
+
+[standalone_forecast_ifs_global21]
+prefix              = standalone
+pgd_source          = static
+surf_ic_source      = static
+# profile (ntasks = ntasks/node)
+time                = 01:00:00
+ntasks              = 32
+nnodes              = 1
+openmp              = 4
+
 
 # MIXS
 # ====
@@ -384,7 +443,27 @@ openmp              = 1
 # ASSIM
 # =====
 
-[BSM_4D_arpege]
+[OOPSBSM_4D_arpege]
+geometry            = geometry(globalupd149)
+timestep            = 1800
+rundate             = 2022012106
+# profile (ntasks = ntasks/node)
+time                = 01:00:00
+ntasks              = 16
+nnodes              = 1
+openmp              = 8
+
+[OOPSBA1_4D_arpege]
+geometry            = geometry(globalupd149)
+timestep            = 1800
+rundate             = 2022012106
+# profile (ntasks = ntasks/node)
+time                = 01:00:00
+ntasks              = 16
+nnodes              = 1
+openmp              = 8
+
+[CNT0BSM_4D_arpege]
 geometry            = geometry(globalupd149)
 timestep            = 1800
 rundate             = 2022012106
