+++ conf/taranis.ini
@@ -188,6 +188,10 @@ timestep            = 90
 geometry            = geometry(antwrp1300)
 timestep            = 60
 
+[chmh2325]
+geometry            = geometry(chmh2325)
+timestep            = 90
+
 # ======================================================================= TASKS
 
 [gitref2pack]
@@ -275,6 +279,10 @@ rundate             = date(2020081800)
 alaro_version       = 1
 rundate             = date(2020081800)
 
+[forecast-alaro1_sfx-chmh2325]
+alaro_version       = 1_sfx
+rundate             = date(2021022000)
+
 # ======================================================================== JOBS
 
 [ciboulai_xpsetup]
@@ -298,7 +306,7 @@ partition           = login
 # FORECASTS
 # =========
 
-[canonical_forecasts]
+[canonical_forecast_arpege_global798c22]
 prefix              = canonical
 # profile (ntasks = ntasks/node)
 time                = 01:00:00
@@ -306,7 +314,15 @@ ntasks              = 32
 nnodes              = 4
 openmp              = 4
 
-[standalone_forecasts]
+[canonical_forecast_arpege_apl_global798c22]
+prefix              = canonical
+# profile (ntasks = ntasks/node)
+time                = 01:00:00
+ntasks              = 32
+nnodes              = 4
+openmp              = 4
+
+[standalone_forecast_alaro0_antwrp1300]
 prefix              = standalone
 pgd_source          = static
 surf_ic_source      = static
@@ -315,6 +331,57 @@ time                = 01:00:00
 ntasks              = 32
 nnodes              = 1
 openmp              = 4
+
+[standalone_forecast_alaro1_antwrp1300]
+prefix              = standalone
+pgd_source          = static
+surf_ic_source      = static
+# profile (ntasks = ntasks/node)
+time                = 01:00:00
+ntasks              = 32
+nnodes              = 1
+openmp              = 4
+
+[standalone_forecast_arome_corsica2500]
+prefix              = standalone
+pgd_source          = static
+surf_ic_source      = static
+# profile (ntasks = ntasks/node)
+time                = 01:00:00
+ntasks              = 32
+nnodes              = 1
+openmp              = 4
+
+[standalone_forecast_arpege_globaltst149c24]
+prefix              = standalone
+pgd_source          = static
+surf_ic_source      = static
+# profile (ntasks = ntasks/node)
+time                = 01:00:00
+ntasks              = 32
+nnodes              = 1
+openmp              = 4
+
+[standalone_forecast_arpege_nosfx_globaltst149c24]
+prefix              = standalone
+pgd_source          = static
+surf_ic_source      = static
+# profile (ntasks = ntasks/node)
+time                = 01:00:00
+ntasks              = 32
+nnodes              = 1
+openmp              = 4
+
+[standalone_forecast_ifs_global21]
+prefix              = standalone
+pgd_source          = static
+surf_ic_source      = static
+# profile (ntasks = ntasks/node)
+time                = 01:00:00
+ntasks              = 32
+nnodes              = 1
+openmp              = 4
+
 
 # MIXS
 # ====
@@ -373,7 +440,27 @@ openmp              = 1
 # ASSIM
 # =====
 
-[BSM_4D_arpege]
+[OOPSBSM_4D_arpege]
+geometry            = geometry(globalupd149)
+timestep            = 1800
+rundate             = 2022012106
+# profile (ntasks = ntasks/node)
+time                = 01:00:00
+ntasks              = 16
+nnodes              = 1
+openmp              = 8
+
+[OOPSBA1_4D_arpege]
+geometry            = geometry(globalupd149)
+timestep            = 1800
+rundate             = 2022012106
+# profile (ntasks = ntasks/node)
+time                = 01:00:00
+ntasks              = 16
+nnodes              = 1
+openmp              = 8
+
+[CNT0BSM_4D_arpege]
 geometry            = geometry(globalupd149)
 timestep            = 1800
 rundate             = 2022012106
