+++ src/tasks/assim/OOPSBA1_4D_arpege.py
@@ -7,9 +7,7 @@ from vortex import toolbox
 from vortex.layout.nodes import Driver, Family, LoopFamily
 
 from .raw2odb.batodb import BatorODB
-from .screenings.screeningCNT0 import Screening as ScreeningCNT0
 from .screenings.screeningOOPS import Screening as ScreeningOOPS
-from .minims.minimCNT0 import Minim as MinimCNT0
 from .minims.minimOOPS import Minim as MinimOOPS
 from .minims.Analyse4dvar import Analyse4dvar
 
@@ -21,15 +19,7 @@ def setup(t, **kw):
                 Family(tag='default_compilation_flavour', ticket=t, nodes=[
                     BatorODB(tag='batodb', ticket=t, **kw),
                     # delayed_fail to let the OOPS family run before raising error
-                    Family(tag='cnt0', ticket=t, on_error='delayed_fail', nodes=[
-                        ScreeningCNT0(tag='screeningCNT0', ticket=t, **kw),
-                        MinimCNT0(tag='minimCNT0', ticket=t, **kw),
-                        ], **kw),
                     Family(tag='oops', ticket=t, nodes=[
-                        Family(tag='seq', ticket=t, on_error='delayed_fail', nodes=[
-                            ScreeningOOPS(tag='screeningOOPS', ticket=t, **kw),
-                            MinimOOPS(tag='minimOOPS', ticket=t, **kw),
-                            ], **kw),
                         Family(tag='allinone', ticket=t, nodes=[
                             Analyse4dvar(tag='Analyse4dvar', ticket=t, **kw),
                             ], **kw),
